requirements.txt
asgiref==3.6.0
autopep8==1.6.0
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==4.0.0
django-db-views==0.1.1
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.11.2
gunicorn==20.1.0
mysqlclient==2.1.0
pycodestyle==2.8.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
whitenoise==6.2.0
